Fsync of blocks is optional (off by default). Config value: FileSync.

+++ block/block.go
@@ -265,10 +265,12 @@ func writeBlockToDisk(ref *Ref, block []byte) error {
 		os.Remove(path)
 		return err
 	}
-	if err := f.Sync(); err != nil {
-		f.Close()
-		os.Remove(path)
-		return err
+	if config.FileSync {
+		if err := f.Sync(); err != nil {
+			f.Close()
+			os.Remove(path)
+			return err
+		}
 	}
 	if err := f.Close(); err != nil {
 		os.Remove(path)
